Address Rick's comments.
