+++ examples/planar_gripper/planar_gripper.cc
@@ -8,6 +8,7 @@
 #include "drake/common/find_resource.h"
 #include "drake/examples/planar_gripper/planar_gripper_common.h"
 #include "drake/examples/planar_gripper/planar_gripper_lcm.h"
+#include "drake/examples/planar_gripper/planar_gripper_utils.h"
 #include "drake/geometry/scene_graph.h"
 #include "drake/multibody/parsing/parser.h"
 #include "drake/multibody/plant/multibody_plant.h"
@@ -228,25 +229,6 @@ void PlanarGripper::SetupPlanarBrick(std::string orientation) {
 void PlanarGripper::SetupPinBrick(std::string orientation) {
   SetupPlant(orientation, "drake/examples/planar_gripper/1dof_brick.sdf");
 }
-
-namespace {
-geometry::GeometryId GetFingertipSphereGeometryId(
-    const multibody::MultibodyPlant<double>& plant,
-    const geometry::SceneGraphInspector<double>& inspector, Finger finger) {
-  return inspector.GetGeometryIdByName(
-      plant.GetBodyFrameIdOrThrow(
-          plant.GetBodyByName(to_string(finger) + "_tip_link").index()),
-      geometry::Role::kProximity, "planar_gripper::tip_sphere_collision");
-}
-
-geometry::GeometryId GetBrickGeometryId(
-    const multibody::MultibodyPlant<double>& plant,
-    const geometry::SceneGraphInspector<double>& inspector) {
-  return inspector.GetGeometryIdByName(
-      plant.GetBodyFrameIdOrThrow(plant.GetBodyByName("brick_link").index()),
-      geometry::Role::kProximity, "brick::box_collision");
-}
-}  // namespace
 
 void PlanarGripper::SetupPlant(std::string orientation,
                                std::string brick_file_name) {
@@ -540,102 +522,6 @@ double PlanarGripper::GetBrickMass() const {
   return dynamic_cast<const multibody::RigidBody<double>&>(
              plant_->GetFrameByName("brick_link").body())
       .default_mass();
-}
-
-namespace {
-std::pair<std::unordered_set<BrickFace>, Eigen::Vector3d>
-GetClosestFacesToFingerImpl(const geometry::SceneGraph<double>& scene_graph,
-                            geometry::GeometryId finger_sphere_geometry_id,
-                            geometry::GeometryId brick_geometry_id,
-                            const systems::InputPort<double>& query_port,
-                            const systems::Context<double>& plant_context) {
-  const auto& inspector = scene_graph.model_inspector();
-
-  const auto& query_object =
-      query_port.template Eval<geometry::QueryObject<double>>(plant_context);
-  const geometry::SignedDistancePair<double> signed_distance_pair =
-      query_object.ComputeSignedDistancePairClosestPoints(
-          finger_sphere_geometry_id, brick_geometry_id);
-  const geometry::Box& box_shape =
-      dynamic_cast<const geometry::Box&>(inspector.GetShape(brick_geometry_id));
-  std::unordered_set<BrickFace> closest_faces;
-  const Eigen::Vector3d p_BCb =
-      inspector.GetPoseInFrame(brick_geometry_id) * signed_distance_pair.p_BCb;
-  if (std::abs(signed_distance_pair.p_BCb(1) - box_shape.depth() / 2) < 1e-3) {
-    closest_faces.insert(BrickFace::kPosY);
-  } else if (std::abs(signed_distance_pair.p_BCb(1) + box_shape.depth() / 2) <
-             1e-3) {
-    closest_faces.insert(BrickFace::kNegY);
-  }
-  if (std::abs(signed_distance_pair.p_BCb(2) - box_shape.height() / 2) < 1e-3) {
-    closest_faces.insert(BrickFace::kPosZ);
-  } else if (std::abs(signed_distance_pair.p_BCb(2) + box_shape.height() / 2) <
-             1e-3) {
-    closest_faces.insert(BrickFace::kNegZ);
-  }
-  return std::make_pair(closest_faces, p_BCb);
-}
-}  // namespace
-
-std::pair<std::unordered_set<BrickFace>, Eigen::Vector3d>
-GetClosestFacesToFinger(const multibody::MultibodyPlant<double>& plant,
-                        const geometry::SceneGraph<double>& scene_graph,
-                        const systems::Context<double>& plant_context,
-                        Finger finger) {
-  const auto& inspector = scene_graph.model_inspector();
-  const geometry::GeometryId finger_sphere_geometry_id =
-      GetFingertipSphereGeometryId(plant, inspector, finger);
-  const geometry::GeometryId brick_geometry_id =
-      GetBrickGeometryId(plant, inspector);
-  const auto& query_port = plant.get_geometry_query_input_port();
-  return GetClosestFacesToFingerImpl(scene_graph, finger_sphere_geometry_id,
-                                     brick_geometry_id, query_port,
-                                     plant_context);
-}
-
-FingerFaceAssigner::FingerFaceAssigner(
-    const multibody::MultibodyPlant<double>& plant,
-    const geometry::SceneGraph<double>& scene_graph)
-    : plant_{plant}, scene_graph_{scene_graph} {
-  const auto& inspector = scene_graph_.model_inspector();
-  for (const auto finger :
-       {Finger::kFinger1, Finger::kFinger2, Finger::kFinger3}) {
-    finger_sphere_geometry_ids_.emplace(
-        finger, GetFingertipSphereGeometryId(plant_, inspector, finger));
-  }
-  DRAKE_DEMAND(static_cast<int>(finger_sphere_geometry_ids_.size()) ==
-               kNumFingers);
-  brick_geometry_id_ = GetBrickGeometryId(plant, inspector);
-  geometry_query_input_port_ =
-      this->DeclareAbstractInputPort("geometry_query",
-                                     Value<geometry::QueryObject<double>>{})
-          .get_index();
-  finger_face_assignments_output_port_ =
-      this->DeclareAbstractOutputPort("finger_face_assignments",
-                                      &FingerFaceAssigner::CalcOutput)
-          .get_index();
-}
-
-void FingerFaceAssigner::CalcOutput(
-    const systems::Context<double>& context,
-    std::unordered_map<Finger, std::pair<BrickFace, Eigen::Vector2d>>*
-        finger_face_assignments) const {
-  finger_face_assignments->clear();
-  const auto& query_port = this->get_geometry_query_input_port();
-  for (const auto& finger_id_pair : finger_sphere_geometry_ids_) {
-    const std::pair<std::unordered_set<BrickFace>, Eigen::Vector3d>
-        finger_faces = GetClosestFacesToFingerImpl(
-            scene_graph_, finger_id_pair.second, brick_geometry_id_, query_port,
-            context);
-    // There must exist at least one closest face.
-    DRAKE_DEMAND(!finger_faces.first.empty());
-    // If there are multiple closest faces (when the witness point is a vertex
-    // of the brick), we arbitrarily choose the first closest face.
-    finger_face_assignments->emplace(
-        finger_id_pair.first,
-        std::make_pair(*finger_faces.first.begin(),
-                       Eigen::Vector2d(finger_faces.second.tail<2>())));
-  }
 }
 
 }  // namespace planar_gripper
